+++ kube_scheduler/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ***Added***:
 
-* Add scheduler_goroutines metric to kube_scheduler check ([#15697](https://github.com/DataDog/integrations-core/pull/15697))
+* Add goroutine_by_scheduling_operation metric to kube_scheduler check ([#15697](https://github.com/DataDog/integrations-core/pull/15697))
 
 ## 4.6.1 / 2023-08-18